backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9

sqlalchemy[asyncio]==2.0.30
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary==2.9.9

python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

redis==5.0.4
hiredis==2.3.2

celery[redis]==5.3.6
kombu==5.3.3

boto3==1.34.110
minio==7.2.7

Pillow==10.3.0
opencv-python-headless==4.9.0.80
pdf2image==1.17.0
pytesseract==0.3.10

anthropic==0.28.0

httpx==0.27.0
sendgrid==6.11.0

pydantic==2.7.1
pydantic-settings==2.2.1
email-validator==2.1.1

cryptography==42.0.7
python-dotenv==1.0.1
python-dateutil==2.9.0
pytz==2024.1

pytest==8.2.0
pytest-asyncio==0.23.6
